pyproject.toml
[project]
name = "ssmd"
version = "0.1.0"
description = "An elegant backend for interaction with ssmd database."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "fastapi>=0.135.2",
  "psycopg2-binary>=2.9.11",
  "pwdlib[argon2]>=0.3.0",
  "pydantic>=2.12.5",
  "python-dotenv>=1.2.2",
  "python-jose[cryptography]>=3.5.0",
  "python-multipart>=0.0.22",
  "sqlalchemy>=2.0.48",
  "uvicorn>=0.42.0",
]
